[!] Full extend DADDU (Fixes cut scene in final stage in Conker) (Note : This is the proper way, not sure why ExciteBike needs the operand to be 32bit.. I'll investigate eventually) [!] Fixed DSUBU (Fixes crash in final stage in Conker, also random lock ups in Animal Crossing, and random lock up when entering pause menu in Majora's Mask)
PS : Final stage in Conker should be playable now \o/

Note: There's a big design flaw in our SPECIAL opcodes (and probably other opcodes), since rt and rs order is wrong! This doesn't matter for mults and adds, but is a break deal for subs and divs, I'll correct this soon.

git-svn-id: https://subversion.assembla.com/svn/Daedalusx64/trunk@330 42e9bfbe-799a-4a2d-bad1-236e862a387a